requirements.txt
# --- 🌐 Web Server & API ---
fastapi>=0.110.0,<1.0.0
uvicorn>=0.29.0
gunicorn>=21.2.0
requests>=2.31.0
httpx>=0.27.0                  # MCP 서버 비동기 HTTP 클라이언트
python-multipart>=0.0.9
slowapi>=0.1.9                 # API 속도 제한 (Rate Limiting)

# --- 🧠 AI & LLM Frameworks ---
langchain>=0.1.13,<0.3.0
langchain-core>=0.1.33,<0.3.0
langchain-community>=0.0.29,<0.3.0
langchain-openai>=0.1.1,<0.3.0
langchain-huggingface>=0.0.1,<0.2.0
langchain-ollama>=0.1.0,<0.3.0
langchain-google-genai>=1.0.1,<2.0.0
langchain-text-splitters>=0.0.1,<0.3.0

# --- 📄 Advanced Document Parsing ---
llama-parse>=0.4.2
llama-index>=0.10.20,<0.11.0
unstructured[all-docs]>=0.13.0,<0.14.0
beautifulsoup4>=4.12.0
lxml>=5.1.0
pypdf>=4.1.0
openpyxl>=3.1.0                # .xlsx 지원
xlrd>=2.0.1                    # .xls (구형 엑셀) 지원
python-magic>=0.4.27           # 파일 타입 감지 (Linux: libmagic1 필요)
pandas>=2.2.0

# --- 🗄️ Database & Vector Store ---
psycopg2-binary>=2.9.9
pgvector>=0.2.5
langchain-postgres>=0.0.9
asyncpg>=0.29.0                # 비동기 PostgreSQL 드라이버

# --- 🎯 Embedding & Reranking ---
sentence-transformers>=2.6.0
torch>=2.2.0 --index-url https://download.pytorch.org/whl/cpu

# --- 🔒 Security & Utilities ---
python-dotenv>=1.0.1
cryptography>=42.0.5
tqdm>=4.66.0
